chore: move router.js into index
